+++ src/Console/Command/BatchScriptCommand.php
@@ -4,6 +4,7 @@ namespace PhpZone\Shell\Console\Command;
 
 use PhpZone\Shell\Exception\Console\Command\InvalidScriptException;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -15,6 +16,9 @@ class BatchScriptCommand extends Command
 
     /** @var bool */
     private $enabled = true;
+
+    /** @var bool */
+    private $tty = true;
 
     /**
      * @param string $name
@@ -40,6 +44,10 @@ class BatchScriptCommand extends Command
             $this->enabled = $options['enable'];
         }
 
+        if (isset($options['tty'])) {
+            $this->tty = $options['tty'];
+        }
+
         parent::__construct($name);
     }
 
@@ -55,7 +63,13 @@ class BatchScriptCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $scriptInput = clone $input;
+        $inputParameters = array(
+            '--no-tty' => !$this->tty
+        );
+
+        if ($input->getOption('no-tty')) {
+            $inputParameters['--no-tty'] = $input->getOption('no-tty');
+        }
 
         foreach ($this->batchScript as $script) {
             $uniqueId = uniqid($this->getName() . ':');
@@ -63,10 +77,10 @@ class BatchScriptCommand extends Command
             $command = new ScriptCommand($uniqueId, $script);
             $this->getApplication()->add($command);
 
-            $scriptInput->setArgument('command', $uniqueId);
+            $inputParameters['command'] = $uniqueId;
 
             $command = $this->getApplication()->find($uniqueId);
-            $command->run($scriptInput, $output);
+            $command->run(new ArrayInput($inputParameters), $output);
         }
     }